fix(error): return u8 from exit_code to match ExitCode::from

+++ src/error.rs
@@ -1,3 +1,4 @@
+// Variants added ahead of their first callers; will be removed once all are constructed.
 #[allow(dead_code)]
 #[derive(Debug, thiserror::Error)]
 pub enum CliError {
@@ -12,7 +13,7 @@ pub enum CliError {
 }
 
 impl CliError {
-    pub fn exit_code(&self) -> i32 {
+    pub fn exit_code(&self) -> u8 {
         match self {
             CliError::Api { .. } | CliError::Network(_) => 1,
             CliError::Usage(_) => 2,
@@ -27,20 +28,20 @@ mod tests {
     #[test]
     fn api_error_exits_1() {
         let err = CliError::Api { code: 250, message: "No search parameters provided.".into() };
-        assert_eq!(err.exit_code(), 1);
+        assert_eq!(err.exit_code(), 1u8);
         assert_eq!(err.to_string(), "Elvanto returned code 250: No search parameters provided.");
     }
 
     #[test]
     fn network_error_exits_1() {
         let err = CliError::Network("connection timed out".into());
-        assert_eq!(err.exit_code(), 1);
+        assert_eq!(err.exit_code(), 1u8);
         assert_eq!(err.to_string(), "network: connection timed out");
     }
 
     #[test]
     fn usage_error_exits_2() {
         let err = CliError::Usage("ELVANTO_API_KEY is not set".into());
-        assert_eq!(err.exit_code(), 2);
+        assert_eq!(err.exit_code(), 2u8);
     }
 }
